ci: refactor build setup (#43)

+++ src/circle_detection/operations/_deduplicate_circles.py
@@ -69,7 +69,10 @@ def deduplicate_circles(
 
     # add batch indices as first dimension to separate circles from different batch items
     rounded_circles = np.empty((len(circles), 4), dtype=circles.dtype)
-    rounded_circles[:, 0] = np.repeat(np.arange(len(batch_lengths), dtype=circles.dtype), batch_lengths)
+
+    # using np.int64 on 32 bit systems throws an error here, so we need to use np.intp
+    # see https://github.com/numpy/numpy/issues/4384
+    rounded_circles[:, 0] = np.repeat(np.arange(len(batch_lengths), dtype=circles.dtype), batch_lengths.astype(np.intp))
     rounded_circles[:, 1:] = np.round(circles, decimals=deduplication_precision)
 
     unique_rounded_circles, selected_indices = np.unique(rounded_circles, return_index=True, axis=0)
